@@ -7,9 +7,9 @@
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
     "normalize.css": "^8.0.1",
-    "react": "^16.14.0",
+    "react": "^19.1.0",
     "react-div-100vh": "^0.5.6",
-    "react-dom": "^16.14.0",
+    "react-dom": "^19.1.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1"
   },
@@ -35,6 +35,8 @@
     ]
   },
   "devDependencies": {
+    "@types/react": "^18.3.23",
+    "@types/react-dom": "^18.3.7",
     "url-loader": "^4.1.1"
   }
 }
